Revert recent Travis CI changes
This reverts commits ac81ba60ea98e878c34c84f67e4486d9f68d5d5b and
519caad6ea868b9c5ead545d7d7a6e7f34888b71, as (a) the Travis CI build
broke, and (b) the JitPack build did not. All we actually need for our
(bisq's) purposes is for the JitPack build to work; if it fails, we'll
know, because we won't be able to consume our latest changes. We can
track down build failures that way.  Having an additional Travis CI
build is probably just confusing.

See bitsquare/bitsquare#898