Fix broadcast bug when broadcastToAllPeers is true
If the broadcastToAllPeers flag is true and we would use all peers for
broadcasting for some unknown reason we don not hear back from the
broadcast or hear back very late (mostly > 1 minute). If we do not
use all peers but at least 1 less it works reliable fast (about 3
seconds).
It is not clear why that happens but is likely related to some
behaviour in other classes for handling the events to hear back from
the nodes.